requirements/test.txt
# Test dependencies go here.
-r base.txt

# Python-PostgreSQL Database Adapter
# If using Win for dev, this assumes Unix in test/prod
psycopg2==2.6.2

coverage==4.3.4
flake8==3.2.1 # pyup: != 2.6.0
django-test-plus==1.0.16
factory-boy==2.8.1

# pytest
pytest-django==3.1.2
pytest-sugar==0.8.0

selenium

model_mommy==1.3.1